backend/VSL-text2sign-model/requirements.txt
absl-py==2.2.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
protobuf==6.30.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.7.0
tqdm==4.67.1
typing_extensions==4.13.2
Werkzeug==3.1.3
